rename: organize landing page components
