+++ client/src/pages/Landing/components/Sponsors/Sponsors.css
